requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
boost-cpp=1.74.0=hff03dee_4
bzip2=1.0.8=h0d85af4_4
c-ares=1.17.2=h0d85af4_0
ca-certificates=2021.10.8=h033912b_0
cairo=1.16.0=he43a7df_1008
certifi=2021.10.8=py39h6e9494a_0
cfitsio=3.470=h01dc385_7
curl=7.79.1=hf45b732_1
cycler=0.10.0=py_2
cython=0.29.24=py39h9fcab8e_0
expat=2.4.1=he49afe7_0
ffmpeg=4.4.0=h4dad6da_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.1=h10f422b_1005
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
freetype=2.10.4=h4cff582_1
freexl=1.0.6=h0d85af4_0
gdal=3.3.2=py39h9e6e899_4
geos=3.9.1=he49afe7_2
geotiff=1.7.0=hb0faedc_3
gettext=0.19.8.1=hd1a6beb_1008
giflib=5.2.1=hbcb3906_2
gmp=6.2.1=h2e338ed_0
gnutls=3.6.13=h756fd2b_1
graphite2=1.3.13=h2e338ed_1001
harfbuzz=3.0.0=h159f659_1
hdf4=4.2.15=hefd3b78_3
hdf5=1.12.1=nompi_h2f0ef1a_101
icu=68.1=h74dc148_0
imglyb=2.0.0=pyh8a188c0_0
jasper=2.0.33=h013e400_0
jbig=2.1=h0d85af4_2003
jgo=1.0.3=pyhd8ed1ab_0
jpeg=9d=hbcb3906_0
jpype1=1.3.0=py39hf018cea_0
json-c=0.15=hcb556a6_0
kealib=1.4.14=ha22a8b1_3
kiwisolver=1.3.2=py39hf018cea_0
krb5=1.19.2=hcfbf3a7_2
lame=3.100=h35c211d_1001
lcms2=2.12=h577c468_0
lerc=3.0=he49afe7_0
libblas=3.9.0=12_osx64_openblas
libcblas=3.9.0=12_osx64_openblas
libcurl=7.79.1=hf45b732_1
libcxx=12.0.1=habf9029_0
libdap4=3.20.6=h3e144a0_2
libdeflate=1.8=h0d85af4_0
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libffi=3.4.2=he49afe7_4
libgdal=3.3.2=h053b6da_4
libgfortran=5.0.0=9_3_0_h6c81a4c_23
libgfortran5=9.3.0=h6c81a4c_23
libglib=2.70.0=hf1fb8c0_0
libiconv=1.16=haf1e3a3_0
libkml=1.3.0=h8fd9edb_1014
liblapack=3.9.0=12_osx64_openblas
liblapacke=3.9.0=12_osx64_openblas
libnetcdf=4.8.1=nompi_h6609ca0_101
libnghttp2=1.43.0=h6f36284_1
libopenblas=0.3.18=openmp_h3351f45_0
libopencv=4.5.3=py39hd2a74c5_5
libpng=1.6.37=h7cec526_2
libpq=13.3=hea3049e_1
libprotobuf=3.18.1=hcf210ce_0
librttopo=1.1.0=h5413771_6
libspatialite=5.0.1=hfc891e7_9
libssh2=1.10.0=h52ee1ee_2
libtiff=4.3.0=hd146c10_2
libwebp-base=1.2.1=h0d85af4_0
libxml2=2.9.12=h93ec3fd_0
libzip=1.8.0=h8b0c345_1
libzlib=1.2.11=h9173be1_1013
llvm-openmp=12.0.1=hda6cdc1_1
lz4-c=1.9.3=he49afe7_1
matplotlib-base=3.4.3=py39hb07454d_1
maven=3.6.3=h694c41f_0
ncurses=6.2=h2e338ed_4
nettle=3.6=hedd7734_0
nspr=4.30=hcd9eead_0
nss=3.69=h31e2bf1_1
numpy=1.21.2=py39h7eed0ac_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.3=py39h6e9494a_5
openh264=2.1.1=hfd3ada9_0
openjdk=8.0.302=h0d85af4_0
openjpeg=2.4.0=h6e7aa92_1
openssl=1.1.1l=h0d85af4_0
pandas=1.3.4=py39h4d6be9b_0
pcre=8.45=he49afe7_0
pillow=8.3.2=py39he9bb72f_0
pip=21.3=pyhd8ed1ab_0
pixman=0.40.0=hbcb3906_0
poppler=21.09.0=h9573804_3
poppler-data=0.4.11=hd8ed1ab_0
postgresql=13.3=he8fe76e_1
proj=8.1.1=h1512c50_2
psutil=5.8.0=py39h89e85a6_1
py-opencv=4.5.3=py39h71a6800_5
pycocotools=2.0.2=py39hc89836e_2
pyimagej=1.0.2=py39h6e9494a_0
pyparsing=2.4.7=pyh9f0ad1d_0
python=3.9.7=h1248fe1_3_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.9=2_cp39
pytz=2021.3=pyhd8ed1ab_0
readline=8.1=h05e3726_0
roifile=2021.6.6=pypi_0
scyjava=1.3.1=pyhd8ed1ab_0
setuptools=58.2.0=py39h6e9494a_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.36.0=h23a322b_2
tifffile=2021.10.12=pypi_0
tiledb=2.3.4=h8370e7a_0
tk=8.6.11=h5dbffcc_1
tornado=6.1=py39h89e85a6_1
tzcode=2021d=h0d85af4_0
tzdata=2021d=he74cb21_0
wheel=0.37.0=pyhd8ed1ab_1
x264=1!161.3030=h0d85af4_1
xarray=0.19.0=pyhd8ed1ab_1
xerces-c=3.2.3=h379762d_3
xz=5.2.5=haf1e3a3_1
zlib=1.2.11=h9173be1_1013
zstd=1.5.0=h582d3a0_0
